batch: Move realized file content
Batch storage owns text publication and contains the helper that
materializes selected ownership into file bytes.

That helper is useful outside persistence when commands need to compare a
stored batch target with live content. Keeping it in storage makes
non-storage planning code depend on the publication module.

This commit begins separating that boundary by moving realized text
construction into a batch.realized_file_content module and updating
production callers.

Subsequent commits will move tests to the new boundary and guard the module
owner.
